backend/requirements.txt
aiodns==3.0.0
aiofile==3.8.1
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
caio==0.9.8
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
Django==4.1
django-localflavor==3.1
djangorestframework==3.13.1
frozenlist==1.3.1
multidict==6.0.2
idna==3.3
psycopg2==2.9.3
pycares==4.2.2
pycparser==2.21
PyJWT==2.4.0
python-dotenv==1.0.0
python-stdnum==1.17
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
stream-chat==4.4.1
urllib3==1.26.11
yarl==1.8.1
